requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16
psycopg2-binary>=2.9
drf-spectacular>=0.27
channels>=4.0
daphne>=4.0
